Fix: main page, memoization
