+++ src/pages/Main/components/Footer/style.jsx
@@ -1,0 +1,40 @@
+import styled from 'styled-components';
+
+export const FooterWrapper = styled.div`
+  background-color: #ffffff;
+  font-family: 'Pretendard';
+  position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 80px;
+  z-index: 1;
+  display: flex;
+  gap: 8px;
+
+  box-shadow: 0px 4px 8px 3px rgba(0, 0, 0, 0.3), 0px 1px 3px rgba(0, 0, 0, 0.1);
+
+  div.menu_container {
+    padding-top: 8px;
+    padding-left: 14px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    cursor: pointer;
+  }
+
+  .menu_border {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    width: 64px;
+    height: 32px;
+    border-radius: 16px;
+  }
+  .menu_text {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 12px;
+    line-height: 18px;
+  }
+`;
