+++ src/pages/Main/components/MainButton/style.jsx
@@ -1,0 +1,23 @@
+import styled from 'styled-components';
+
+export const MainButtonWrapper = styled.button`
+  position: absolute;
+  width: 56px;
+  height: 56px;
+  border-radius: 50%;
+  background: #00d082;
+  bottom: 100px;
+  right: 17px;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  cursor: pointer;
+  border: none;
+
+  img {
+    width: 18px;
+    height: 18px;
+  }
+`;
